+++ src/editor/core/draw/particle/previewer/Previewer.ts
@@ -172,12 +172,10 @@ export class Previewer {
         if (this.curElement) {
           this.curElement.width = this.width
           this.curElement.height = this.height
-          this.draw.render({ isSetCursor: false })
-          this.drawResizer(
-            this.curElement,
-            this.curPosition,
-            this.previewerDrawOption
-          )
+          this.draw.render({
+            isSetCursor: true,
+            curIndex: this.curPosition?.index
+          })
         }
         // 还原副作用
         this.resizerImageContainer.style.display = 'none'
@@ -423,7 +421,19 @@ export class Previewer {
     position: IElementPosition | null = null,
     options: IPreviewerDrawOption = {}
   ) {
+    // 缓存配置
     this.previewerDrawOption = options
+    this.curElementSrc = element[options.srcKey || 'value'] || ''
+    // 更新渲染尺寸及位置
+    this.updateResizer(element, position)
+    // 监听事件
+    document.addEventListener('keydown', this._keydown)
+  }
+
+  public updateResizer(
+    element: IElement,
+    position: IElementPosition | null = null
+  ) {
     const { scale } = this.options
     const elementWidth = element.width! * scale
     const elementHeight = element.height! * scale
@@ -439,12 +449,11 @@ export class Previewer {
     // 更新预览包围框尺寸
     this._updateResizerRect(elementWidth, elementHeight)
     this.resizerSelection.style.display = 'block'
+    // 缓存基础信息
     this.curElement = element
-    this.curElementSrc = element[options.srcKey || 'value'] || ''
     this.curPosition = position
-    this.width = this.curElement.width! * scale
-    this.height = this.curElement.height! * scale
-    document.addEventListener('keydown', this._keydown)
+    this.width = elementWidth
+    this.height = elementHeight
   }
 
   public clearResizer() {