+++ src/editor/interface/table/Table.ts
@@ -1,0 +1,7 @@
+import { IPadding } from '../Common'
+
+export interface ITableOption {
+  tdPadding?: IPadding
+  defaultTrMinHeight?: number
+  defaultColMinWidth?: number
+}