+++ src/editor/interface/table/Tr.ts
@@ -5,4 +5,5 @@ export interface ITr {
   height: number
   tdList: ITd[]
   minHeight?: number
+  pagingRepeat?: boolean // 在各页顶端以标题行的形式重复出现
 }